# openapi.json
{
    "openapi": "3.1.0",
    "info": { "title": "FastAPI", "version": "0.1.0" },
    "paths": {
        "/api/circle": {
            "post": {
                "summary": "Api Circle",
                "description": "Trigger the rainbow circle.",
                "operationId": "api_circle_api_circle_post",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": {} } }
                    }
                }
            }
        },
        "/api/flash-direction": {
            "post": {
                "summary": "Api Flash",
                "description": "Flashes the LED strip in a given direction.\n\nArgs:\n    direction (int): 0: bottom, 1: left, 2: top, 3: right",
                "operationId": "api_flash_api_flash_direction_post",
                "parameters": [
                    {
                        "name": "direction",
                        "in": "query",
                        "required": true,
                        "schema": { "type": "integer", "title": "Direction" }
                    },
                    {
                        "name": "number_of_flashes",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Number Of Flashes"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": {} } }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/pulse": {
            "post": {
                "summary": "Api Pulse",
                "description": "Pulses the LED strip between given colours.",
                "operationId": "api_pulse_api_pulse_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PulseRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": {} } }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/clear": {
            "post": {
                "summary": "Api Clear",
                "description": "Clears the strip.",
                "operationId": "api_clear_api_clear_post",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": {} } }
                    }
                }
            }
        },
        "/api/solid": {
            "post": {
                "summary": "Api Solid",
                "description": "Sets the LED strip to a solid colour.",
                "operationId": "api_solid_api_solid_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SolidRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": {} } }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/percentage": {
            "post": {
                "summary": "Api Circle",
                "description": "Set the LED to a red-amber-green based percentage.",
                "operationId": "api_circle_api_percentage_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PercentageRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": { "application/json": { "schema": {} } }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "PercentageRequest": {
                "properties": {
                    "percentage": { "type": "number", "title": "Percentage" },
                    "flashing": {
                        "type": "boolean",
                        "title": "Flashing",
                        "default": true
                    }
                },
                "type": "object",
                "required": ["percentage"],
                "title": "PercentageRequest"
            },
            "PulseRequest": {
                "properties": {
                    "colours": {
                        "items": { "type": "string" },
                        "type": "array",
                        "title": "Colours",
                        "description": "A list of hex codes for the colours to pulse through"
                    },
                    "pause_time_seconds": {
                        "type": "number",
                        "title": "Pause Time Seconds",
                        "default": 1
                    }
                },
                "type": "object",
                "required": ["colours"],
                "title": "PulseRequest"
            },
            "SolidRequest": {
                "properties": {
                    "hex_code": {
                        "type": "string",
                        "title": "Hex Code",
                        "description": "A list of hex codes for the colours to pulse through",
                        "examples": ["#FF0000", "#00FF00", "#0000FF"]
                    }
                },
                "type": "object",
                "required": ["hex_code"],
                "title": "SolidRequest"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                { "type": "string" },
                                { "type": "integer" }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": { "type": "string", "title": "Message" },
                    "type": { "type": "string", "title": "Error Type" }
                },
                "type": "object",
                "required": ["loc", "msg", "type"],
                "title": "ValidationError"
            }
        }
    }
}